+++ MaxMind.MinFraud.UnitTest/Request/EmailTest.cs
@@ -165,6 +165,10 @@ namespace MaxMind.MinFraud.UnitTest.Request
             e = new Email(address: "foo@example.cam", hashAddress: true);
             Assert.Equal("b48def645758b95537d4424c84d1a9ff", e.AddressMD5);
             Assert.Equal("example.cam", e.Domain);
+
+            e = new Email(address: "foo@10000gmail.com", hashAddress: true);
+            Assert.Equal("6c0fbec2cc554c35c3d2b8b51840b49d", e.AddressMD5);
+            Assert.Equal("10000gmail.com", e.Domain);
         }
 
         [Fact]