+++ MaxMind.MinFraud/Request/Email.cs
@@ -18,6 +18,7 @@ namespace MaxMind.MinFraud.Request
         private static readonly Regex DuplicateDotComRe = new Regex(@"(?:\.com){2,}$", RegexOptions.Compiled);
         private static readonly Regex DotComExtraCharsRe = new Regex(@"\.com[^.]+$", RegexOptions.Compiled);
         private static readonly Regex DotComTypoRe = new Regex(@"(?:\.(?:com|c[a-z]{1,2}m|co[ln]|[dsvx]o[mn]|))$", RegexOptions.Compiled);
+        private static readonly Regex GmailLeadingDigitRe = new Regex(@"^\d+(?:gmail?\.com)$", RegexOptions.Compiled);
         private static readonly IdnMapping _idn = new();
         private static readonly IReadOnlyDictionary<string, string> _typoDomains = new Dictionary<string, string>
         {
@@ -400,6 +401,7 @@ namespace MaxMind.MinFraud.Request
             domain = DuplicateDotComRe.Replace(domain, ".com");
             domain = DotComExtraCharsRe.Replace(domain, ".com");
             domain = DotComTypoRe.Replace(domain, ".com");
+            domain = GmailLeadingDigitRe.Replace(domain, "gmail.com");
 
             if (_typoDomains.ContainsKey(domain))
             {